Calculate chi online and compare to offline calculation
In the previous commit, we found that the approach to calculating chi offline was flawed because
the boundary conditions were not being correctly applied by the derivative. Here, we demonstrate
this by additionally calculating chi online and showing that the online calculation asymptotes
to the analytical equilibrium solution, as expected.