energyflux notebook starts to compute the energy fluxes for topographic HC --> realized we may need to go back to no topography to accurately compute these fluxes. I may not have much time this quarter, so this is where I'll start. Issue #31 on gitub
Also Kolmogorov scale notebook is updated using broadcasting and all function to see if kolmogorov scale is resolved throughout. Found that it's only resolved for first timestep. Then, the theoretical constraint for lambda is used to resolve kolomogorov scale and new resolution is implemented. I want to create few analysis plots for new resolution. Issue #20 on github

this is the HPC3 branch, I'm going to merge both laptop (upgrade-Oceananigans) branch and this one, so everything is consolidated.